+++ src/components/Topic.tsx
@@ -16,20 +16,14 @@ const Topic = (props: TopicProps) => {
     console.log(design);
 
     const handleClick = (title: string) => {
-        if (title !== "new_topic") 
-        {
-            router.push(`/subtopics?topic=${encodeURIComponent(title)}`);
-        } 
-        else 
-        {
-            router.push(`/${title}`);
-        }
+        router.push(`/${title}`);
     };
+
     return (
-    <div className="bg-white rounded-2xl shadow-md hover:shadow-lg transition-shadow duration-300 p-4 flex flex-col items-center space-y-4 w-60">
+        <div className="bg-white rounded-2xl shadow-md hover:shadow-lg transition-shadow duration-300 p-4 flex flex-col items-center space-y-4 w-60">
             <button
                 className="hover:cursor-pointer"
-                onClick={() => handleClick(title) }
+                onClick={() => handleClick(title)}
             >
                 <Image
                     src={
@@ -44,7 +38,9 @@ const Topic = (props: TopicProps) => {
                     alt={`${title} image`}
                     className="rounded-lg"
                 ></Image>
-                <h1 className={`mt-2 text-center text-lg font-semibold text-gray-700 ${colour}`}>
+                <h1
+                    className={`mt-2 text-center text-lg font-semibold text-gray-700 ${colour}`}
+                >
                     {title.replace("_", " ")}
                 </h1>
             </button>
